you can change game speed using a slider

+++ src/game_player/view/IGameView.java
@@ -17,6 +17,6 @@ public interface IGameView {
     
     void changeColorScheme (int colorIndex);
       
-    void changeGameSpeed (int gameSpeed); 
+    void changeGameSpeed (double gameSpeed); 
     
 }
